# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Employee API",
    "description": "This is a Employee API created with Express.js. You can find more about it [here](https://github.com/rameshknr/EmployeeDetails).",
    "contact": {
      "email": "ramesh.navamani@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Login a user",
        "operationId": "loginUser",
        "requestBody": {
          "description": "Login user object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User login successfully"
          },
          "400": {
            "description": "Invalid username and password supplied"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "login": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
    }
  }
}